+++ packages/terra-site/CHANGELOG.md
@@ -7,6 +7,10 @@ Unreleased
 * Updated Avatar import in the code examples
 * Updated the Overlay examples to use non-transpiled code
 
+### Added
+* Examples for terra-form-select props and variants.
+
+
 2.11.0 - (May 16, 2018)
 ------------------
 ### Changed